+++ changelog.md
@@ -1,9 +1,10 @@
 # changelog
 
-## 1.8.3
+## 1.9.0
 * `NEW` specify path of `log` and `meta` by `--logpath=xxx` and `--metapath=XXX` in command line
 * `NEW` completion: worksapce word
 * `NEW` completion: show words in comment
+* `NEW` completion: generate function documentation
 * `CHG` got arg after script name: `lua-language-server.exe main.lua --logpath=D:\log --metapath=D:\meta --develop=false`
 * `FIX` runtime errors